pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "abcd-cluster-phewas"
version = "0.1.0"
description = "Cluster-based PheWAS pipeline for ABCD puberty trajectory analyses"
requires-python = ">=3.10"
dependencies = [
    "pymer4>=0.8,<0.9",
    "pandas>=2.0",
    "numpy>=1.24",
    "scipy>=1.10",
    "statsmodels>=0.14",
    "matplotlib>=3.7",
    "adjustText>=0.8",
    "openpyxl>=3.1",
    "pyyaml>=6.0",
]

[project.scripts]
abcd-phewas = "python_pipeline.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["python_pipeline*"]
